requirements.txt
flask
python-dotenv
psycopg2-binary
PyJWT flask
bcrypt